+++ lib/camera_filter.dart
@@ -1,0 +1,4 @@
+class CameraFilter {
+  static final String NATURAL = "NATURAL";
+  static final String LINE_DRAWING = "LINE_DRAWING";
+}