+++ lib/passive_face_liveness.dart
@@ -15,11 +15,22 @@ class PassiveFaceLiveness {
   String? peopleId;
 
   String? stage;
+  String? filter;
+  bool? enableScreenshot;
 
   PassiveFaceLiveness({required this.mobileToken, this.peopleId});
 
   void setStage(String stage) {
     this.stage = stage;
+  }
+
+  void setCameraFilter(String filter) {
+    this.filter = filter;
+  }
+
+  /// This feature works only for Android right now
+  void setEnableScreenshots(bool enable) {
+    this.enableScreenshot = enable;
   }
 
   Future<PassiveFaceLivenessResult> start() async {
@@ -28,17 +39,19 @@ class PassiveFaceLiveness {
     params["mobileToken"] = mobileToken;
     params["personId"] = peopleId;
     params["stage"] = stage;
+    params["filter"] = filter;
+    params["enableScreenshot"] = enableScreenshot;
 
     Map<dynamic, dynamic> resultMap =
         await _channel.invokeMethod<Map<dynamic, dynamic>>('start', params)
             as Map<dynamic, dynamic>;
 
     bool? success = resultMap["success"];
+
     if (success == null) {
       return new PassiveFaceLivenessClosed();
     } else if (success == true) {
-      return new PassiveFaceLivenessSuccess(
-          signedResponse: resultMap["signedResponse"]);
+      return new PassiveFaceLivenessSuccess(resultMap["signedResponse"]);
     } else {
       return new PassiveFaceLivenessFailure(resultMap["errorMessage"]);
     }