+++ CHANGELOG.md
@@ -1,15 +1,23 @@
-#### 02 de agosto de 2023 
+#### August, 18 2023
+*  **1.2.0**
+  * **new**: New method .setCameraFilter(String filter) to change the SDK camera filter type - Android & iOS
+  * **new**: New method .setEnableScreenshots(bool enable) to enable screenshots during the SDK scan - Android
+
+#### August, 02 2023 
 *  **1.1.2**
   * **Fix**: Fix Documentation erros.
 
-#### 02 de agosto de 2023 
+#### August, 02 2023
+02 de agosto de 2023 
 *  **1.1.1**
   * **Fix**: Fix on the android return of the SDK, it was returning null results in some scenarios.
 
-#### 17 de julho de 2023 
+#### July, 17 2023
+17 de julho de 2023 
 *  **1.1.0**
   * **Mudança**: Now the SDK's return a signed respose in case of success, you must access the parameter `signedResponse` in the `PassiveFaceLivenessSuccess` object.
 
-#### 03 de julho de 2023 
+#### July, 03 2023
+03 de julho de 2023 
 *  **1.0.0**
   * **Novo**: Hello World!!, first publish version.